changed entire project to typescript
