+++ src/Toast/toast.module.css
@@ -71,7 +71,8 @@ body {
   transform: translateX(-50%);
 }
 
-.toast.success {
+.toast.success,
+.toast.default {
   --background-color: #d4edda;
   --border-color: #c3e6cb;
   --text-color: #155724;
